Improve twitter api use case
